Enable relevant tslint rules (part 1) (#94)
## Description

Enable the rules described in #87, with the following exceptions.

These rules will be enabled in a separate changeset, to make them easier to review:

* `promise-function-async: true`
* `no-floating-promises: true`
* `no-unsafe-any: true`

This rule will be enabled once we have a logging framework in place:

* `no-console: true`

## Motivation and Context

<!--- Why is this change required? What problem does it solve? -->
The AWS Toolkit for VS Code is a young product. Enforcing good coding practices now is cheap, and will prevent costly technical debt in the long run. Rules help us in the following areas:

* Functionality: TypeScript and JavaScript have quirks that can lead to unexpected behavior if used unintentionally.
* Maintainability: The code base should be easy to modify without side effects.
* Style: A consistent style makes it easier to read and understand the code base.

## Related Issue(s)

#87 

## Testing

I ran the extension tests from VS Code on Windows 10.

## Checklist

<!--- Go over all the following points, and put an `x` in all the boxes that apply -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're here to help! -->

- [x] I have read the **README** document
- [x] I have read the **CONTRIBUTING** document
- [x] My code follows the code style of this project
- [N/A] I have added tests to cover my changes
- [x] All new and existing tests passed
- [N/A] A short description of the change has been added to the **CHANGELOG**

## License

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.