Update README.md
The instructions under 'Method Two: Create a profile using the AWS Tools for PowerShell ' won't work due to the fact that the credentials are encrypted and VSCode can't get to them. This change is just to update the documentation to remove Method Two and move Method Three into its place.