Wait for debug port to open before trying to attach debugger (#205)
After the sam cli is called to start running a lambda function locally, a check has been added that waits for the debug port to open. This check has been given a configurable max timeout window. Once the debug port is detected, an attempt is made to connect the VS Code debugger to the local lambda function.

Additional change:
* Moved debug port code to reduce non-null assertions