Move the ProfileToolkitCredentialsProvider to jetbrains-core in prep for refresh support
We need to provide a better UX so we need access to the JetBrains APIs so we can provide notifications and actions to remedy issues. This means we need to move the profile support into jetbrains-core. If we ever need to share this we can extract out an abstract ProfileToolkitCredentialsProvider base and move that back