Filter the SAM Runtimes list for Rider (#1330)
Issue: Rider supports Node, our Intellij plugin supports node, but our Rider plugin does not support Node. So, Node shows up in the list of runtimes but generates a C# project.

To fix this, add a filter to SamInitSelectionPanel which is where the runtime list is actually generated so that we only list dotnet runtimes.