Query Debug Adapter for Python Local Debug; Improve Python Test Stability (#866)
Python related integration tests frequently fail so they were disabled. This change makes improvements to Local Debugging for Python SAM Applications, and re-enables the Python related integration tests.

The tests seem to fail because the toolkit is expecting to see text from the sam cli output indicating it is waiting for the debugger to attach, however this output is not getting flushed, leading to a timeout. This has been observed very infrequently in the toolkit on Linux/Mac/Windows.

This change tries something different for Python SAM Applications. Currently, the toolkit uses output to signal when the SAM Application is ready for a debugger to be attached. In this change, we'll try to get a response from the Debug Adapter that the debugger will connect to. Other runtimes behavior remains as-is, and we can evaluate this approach before considering changes to other runtimes.

Integration Testing Changes:
* load the python docker images before running the tests
* activate the python extension at the start of the tests

Out of scope:
* stability improvements to SAM Integration tests for other runtimes