Credentials Management (#888)
This change introduces a more managed approach to implementing Credentials Providers for the toolkit. The goal is to make it easier to onboard support for new credentials types moving forward. This design is very similar to the approach taken in the AWS Toolkit for JetBrains.

Integrating this new system brought about some functional changes and bug fixes.

New Behavior:
* added logging for Shared Credentials profiles detection and validation
* credentials are now shown by "credential provider id" instead of profile name (example: `profile:default`)
* whenever users run the "Connect to AWS" command, the Toolkit's list of available credentials providers is refreshed (eg: pulling the latest updates from the shared credentials files)
* Credentials are cached and re-used for the duration of a toolkit session. When logging in, if a Shared Credentials Profile has been modified since it was cached, the updated version of the profile will be used instead of the cached version now.
* updated the "Invalid Credentials" notification, and added a button guiding users to the logs