Using portfinder instead of static port (#959)
Using node-portfinder instead of static port 5858 when starting the SAM debugger to increase likelihood of avoiding a clash. node-portfinder has been overriden to start searching for ports at port 5858 and will search ports through 65535, using the first available.

OF NOTE: This now produces a Windows Firewall alert for network access on the first debug attempt. Not sure if this affects other OSes.