test: run globalSetup when running single test
`globalSetup` is required for various tests. Include it when running
a single test file (i.e. `$TEST_FILE` is defined via the "Extension
Tests (current file)" task in `launch.json`).

Note: to confirm that the before() and beforeEach() in globalSetup are
executed before any tests, create a test sibling to globalSetup.test.ts,
and observe that the globalSetup before/beforeEach are executed before
executing the "AAA" test:

    src/test/aaaa.test.ts:
        import { getLogger } from '../shared/logger'
        beforeEach(async function() {
            console.log('aaaaa BEFOREEACH')
            // Set every test up so that TestLogger is the logger used by toolkit code
        })
        describe('AAA', async () => {
            it('AAA test', async () => {
                getLogger()
            })
        })