Fix Some Deprecations (#1674)
While looking if the ERROR - debugger.pydev.AbstractCommand - No connection (command: 501 ) errors broke python debugging in 2020.1 (they didn't seem to, I can debug just fine), I noticed that the base getId is now deprecated. Overwriting getId with the default behavior from previous versions. The API is available since 173, so we are good to do it.

Also fixed a raw cast warning and added a comment about when we can fix a deprecation.