Merge private/staging to 'feature/c9toolkit' (#83)
* test: fix deprecated API usage #1001

(node:719) DeprecationWarning: useColors is DEPRECATED and will be removed from a future version of Mocha. Instead, use the "color" option

* update dependencies #1000

* test: run globalSetup when running single test #1003

`globalSetup` is required for various tests. Include it when running
a single test file (i.e. `$TEST_FILE` is defined via the "Extension
Tests (current file)" task in `launch.json`).

Note: to confirm that the before() and beforeEach() in globalSetup are
executed before any tests, create a test sibling to globalSetup.test.ts,
and observe that the globalSetup before/beforeEach are executed before
executing the "AAA" test:

    src/test/aaaa.test.ts:
        import { getLogger } from '../shared/logger'
        beforeEach(async function() {
            console.log('aaaaa BEFOREEACH')
            // Set every test up so that TestLogger is the logger used by toolkit code
        })
        describe('AAA', async () => {
            it('AAA test', async () => {
                getLogger()
            })
        })

* telemetry: Add result metric when submitting feedback #1002

* Adding snippets for aws-sam debug configurations to aid auto-complete (#998)

Co-authored-by: Justin M. Keyes <jmkeyes@amazon.com>
Co-authored-by: Hunter Werlla <werlla@amazon.com>
Co-authored-by: Richard Li <742829+rli@users.noreply.github.com>
Co-authored-by: Kyle Thomson <kylthoms@amazon.com>