Fix Possible SAM Init Race Condition (#1713)
The SAM panel did asynchronous validation of the SAM executable that was not tied to advancement. So, it was possible to press finish and have the executable validation not done yet which would fail even if the SAM cli was valid. Show a dialog while it is validating so this is not possible anymore.