Change Telemetry from three-state to binary (#1048)
The toolkit had a three-state "enable telemetry" setting (enable, disable, defer to VS Code's settings). This has some drawbacks (like complexity in honoring each state -- recently the setting's enum value was corrupted. It is also not so standard -- other extensions use a boolean setting). This change simplifies the toolkit's telemetry setting to a boolean enabled/disabled arrangement.

This is conceptually a reboot of the telemetry setting. Users will be given a new notification on first launch after this change.

If a user was detected as having the old "Disable" value, telemetry will be disabled. All other values (real or fictitious) are treated as enabled.