Create reusable fakes/stubs for VSCode UI components
These components are meant to be injected and used in place of the functions
in the vscode namespace.

Default implementations are provided that simply delegate to vscode.

Fake implementations capture and/or control the outputs
to verify hard-to-test situations without needing separate UI tests.

e.g.

- Verifying progress bars progression
- Verifying messaging on input box validation failures
- Faking commands that would update the UI

They also allow for capturing events that modify global state,
such as updating the system clipboard.

This implementation only covers a few portions of the vscode APIs.
It can be incrementally built out as functionality is needed.