Disable test testParameters_SingleParameterAmazonEvent_Detected (#1759)
this test fails frequently because it gets a: "com.jetbrains.rdclient.util.BackendException: Running number of exploration fell below 0" just suppressing this test would probably make another one fail with the same error. It seems to be caused by a bug in the IDE that happens most on 2019.3 Windows. 

We should investigate this in the future