Add S3 nodes to AWS Explorer tree (#1079)
The following nodes are added:
S3 Nodes (per region)

    Inline: "Create Bucket"
    Context Menu: "Create Bucket"

Bucket Nodes

    Inline: "Create Folder", "Upload File"
    Context Menu: "Upload File" | "Create Folder" | "Copy Name", "Copy Arn"

Folder Nodes

    Inline: "Create Folder", "Upload File"
    Context Menu: "Upload File" | "Create Folder" | "Copy Name", "Copy Arn", "Copy Path"

File Nodes

    Inline: "Download File As"
    Context Menu: "Download File As" | "Copy Name", "Copy Arn", "Copy Path"

More Results Nodes

    Command: "Load More"

Bucket and Folder Nodes cache their children and support pagination.

The very first call to getChildren() always loads and returns the initial page of children.
Subsequent calls to getChildren() always load from the cache.

A call to loadMoreChildren() appends the next page of children to the cache.
The node must subsequently be refreshed to sync the view with the cache.

A full refresh on a Bucket or Folder Node is done by a call to clearChildren().
The nodes then refresh their state and behave as if they had just been created.