YAML Handling Improvements (#1159)
PR turned out larger than I'd hoped but it covers significant ground:
* We no longer produce a minimal template for debugs targeting a template file.
* We still produce a temporary template, but it:
  * Uses all the fields from the old one aside from overwriting the handler name of the targeted function with any refs to parameters or parameter overrides
    * This is to accommodate the Python debug case, which requires a temporary shim file
  * The temporary template is now created as a sibling to the existing template file so the build can use relative pathing for any overrides
* Any parameters in the template file are now override-able through the `sam.template.parameters` field in the launch config.
* We now support `AWS::Lambda::Function`-type resources

As a side-effect, other template-related things (e.g. Lambda layers) should work after doing this. I haven't manually tested this yet.

We may want to consider how we handle the runtime, and handle it like the handler. As of now, if a runtime isn't explicitly provided (read: if it's provided as a ref to a parameter), the template won't validate. This is covered in a comment.