Create SAM App makes sure that the created template.yaml file is registered before creating a launch config (#1401)
Concurrency issue: depending on how Node.js/the extension host/the compute power backing the IDE is feeling, there is no guarantee that the template registry updates in time for it to be used by a workflow. This adds a short 5 second polling period to ensure that the template.yaml file created by the Create SAM App workflow is indeed created before proceeding.