SAM debug/python: drop "heartbeat" concept
Not needed after 65bfa7288f98 and 4471ffd7fe5d. Possibly a "heartbeat"
use-case is to avoid the modal dialog, but that will only happen for
applications that take longer than 1s to start.

Motivation for the heartbeat as described in 77e46e6bea99 mostly was
integration test stability; but that was addressed by various other
changes to timers/retries/tests.