SAM debug: remove "max retries" user setting
This setting is stored as 30 for all users, which is way too high.
Retrying more than a few times is very bad user experience.
Until recently, the retry logic didn't work, so users have never
actually encountered a need for this setting. Expecting them to
configure it will just be confusing, and there is little reason for it
to be configurable anyway.