Speed up integration tests by 20 min (#2283)
In integration tests, tests used to block for 1 minute per test that uses the debugger because we run integration tests under edt. 

In real execution, there's some funky thread switching that leads this call to not be on edt, but that is not emulated in tests. So, skip this entirely if we are in unit test mode. 

Tests have their own timeout which will prevent it running forever when attaching fails.