Fix python and nodejs breakpoints sometimes not working (#2320)
Because path mappings are first one wins, the order matters. The build path was moved before the source path which would break path mappings if the VFS updated the build folder by the time the debugger attached (since build would be in the VFS therefore mapped). In this scenario, it would map to the build folder and not break on breakpoints in the source folder which is undesired.

Additionally, adjust tests to check order of path mappings so this cannot be broken again.