"Create SAM app": default to a more intuitive name
Problem:
Before this commit the "AWS: Create new SAM application" command
defaults the application folder name based on the parent folder:
https://github.com/aws/aws-toolkit-vscode/blob/44dad848fbed20e3b081dfa4cad6f41e00e05a0c/src/lambda/wizards/samInitWizard.ts#L445
In cloud9 this always is `environment`, which is not helpful.

Solution:
Default to a name like "lambda-python3.8", and append a number if such
a name already exists in the workspace.