SSO: fix client registration date calculation #1515
Description
The expiry for the Client Registration in the SSO login flow was being
calculated incorrectly. The expiry in the response object is in seconds and is
now converted to ms, which is the correct value needed to make a new Date
value.

Motivation and Context
This would cause a new client registration every time a user would login
because the expiration date would always be in the past.