fix: Set container environment variables when invoking SAM API (#1631)
## Problem
Debugger was not attaching when using image lambdas in start-api mode.

This was because container environment variables were not being given to SAM CLI. SAM should be setting defaults if we fail to give it environment variables, but for some reason it isn't. It's difficult to say what the exact cause is due to so many parameters being exchanged. One possibility is that something we're setting is preventing the defaults from being used.

## Solution
Add a containerEnvFile field to buildSamCliStartApiArguments.

Fixes #1443