fix: Add initial launch config after 'Create Lambda SAM Application'
## Problem

Initial launch configs were not being added because the 'target' file was changed to 'README.md'.

## Solution

Changed the behavior of addInitialLaunchConfiguration to include templates that reside within the same directory of the target file. The getMainUri function was also exposed so it could be tested. Some additional tests were added.