logger: Fix 'getLogById' #1652
## Problem
The function getLogById would fail to find the recorded log on Windows machines. This was because the file scheme was hardcoded as 'file://', but on Windows local files use the scheme 'file:///'.

## Solution
Require log tracking functions to use vscode.Uri instead of a raw string. This way we do not have to worry about parsing for different operating systems.