credentials: rework semantics of CredentialsProviderId
The first field (formerly "credentialType") is a credential _source_.
The credential _type_ is different.

Both have analogues in `src/shared/telemetry/telemetry.gen.ts`:

    export type CredentialSourceId = 'sharedCredentials' | 'sdkStore' | 'ec2' | 'envVars' | 'other'
    export type CredentialType = 'staticProfile' | 'staticSessionProfile' …

but we can't use the telemetry types because they are too granular.
For storing a CredentialsProviderId (AKA a "fully-qualified
credential"), the source is almost always "profile". See the docstring
for more discussion.

- Introduce CredentialsProviderType.
- Introduce getTelemetryType() and provide a mapping to it from CredentialsProviderType.
- Rename credentialType() to credentialSource().

ref 5785aa819304 #1691