credentials: validate IAM role for EC2 instance creds
Problem:
EC2 instance credentials will fail to resolve if no IAM role is attached to the instance. We are currently not validating this status with the EC2 Metadata Service which gives users the false impression they can use instance credentials.

Solution:
Validate that valid IAM role is attached to the instance before presenting the EC2 profile as an option to the user. This is accomplished by calling the IAM metadata endpoint of the metadata service.