UI: state machine controller (wizard framework 1/3) #1825
State machine controller that will be used in later PRs for the new wizard
framework. Will eventually replace `multiStepWizard.ts` and related tests. The
interface is different enough that it cannot replace it directly.

New capabilities:
* States are remembered and handled exclusively by the controller. This
* prevents mutation of state outside transitions.
* Multiple steps can be added to the machine at runtime as a group, allowing
* for dynamic branching.
* Decouples transitions from the state management logic. 
* Exposes the current step and total number of steps that the controller
* currently has.

### How does this work?
The state controller can be thought of as a higher-order state machine (a
pushdown automaton) that can inject any number of new steps into its sequence
at runtime. Initially, the controller will start with whatever steps are added
through `addStep`. For example, say we add steps `A`, `B` and `C`, then the
flow will look like this:

   A -> B -> C 

Each step is just a function that determines both the next state and whatever
extra steps the controller should add. In this example, let's say step `B`
decides steps `D` and `E` should be added:

   A -> B          C 
          \       /
            D -> E

Of course, if we go back a step from `D` then that branch will cease to exist,
reverting back to the original flow. Step `B` is then processed once more,
deciding what the next state and steps are. Since every step can create new
steps, sophisticated flows can arise without compromising the resulting state.
Note that in the above example, the steps should be thought of more as the
_transitions_ between states, while the actual states are the arrows (initial
and final state not included)

### How does this fit with other pieces?
The wizard controller will use this class to handle execution of steps and keep
track of the proper state. Wizard steps encapsulate the current state plus an
external input from a `Prompter` to determine the next state (assignment of
a field) and next steps (what can we show given the new state?). The public
'step' fields will also be passed to `Prompter` objects so they can process
them accordingly.

### Why?
Managing state in Javascript, and especially with UI, is very difficult since
everything is an object. Having a separate class handling it makes life much
simpler and avoids side effects.