credentials: improved experience for profile creation (#1869)
Problem:
Users routinely complain of difficulty in setting up credentials profile in the Toolkit. The process requires some knowledge of how to create access keys and a guided flow for providing them to the toolkit.
Solution:
This change contains several UX improvements to walk users through the process of setting up a credentials profile and providing them quick access to documentation resources. It also fixes an issue where profiles would not be automatically connected once created.