extensions: start Git extension API wrapper #1900
Wraps some of the Git API functionality in a new class. A wrapper is used since it appears some of the API calls are lacking in their capability, thus we sometimes want to use git directly. This wrapper always 'silently' (but with logs) fails when the extension does not exist or is not enabled. The idea being that we do not want to be strictly dependent on other extensions; they should enhance our own functionality. Of course, a caller can always check for enablement of the wrapper since there are definitely circumstances where behavior should be determined based off what extensions are available.