UI: prompter implementations (wizard framework 3/3) #1874
Follow-up to #1867 (read this PR for detailed information on the framework)

## Overview

These are the concrete implementations of the Prompter abstraction introduced in the previous PR.
For the most part, they do not differ much with their predecessors (that is, they have all the same functionality).

Additionally, a new QuickPick (called CustomQuickPickPrompter) was introduced that generalizes the logic of allowing 'custom' input. Consider a prompt that shows the user a list of Github repositories. Normally, the input for a QuickPick just filters this list. With this new QuickPick, the filter box will filter the list and create a new option for the user to pick from (from the Git extension):
Screen Shot 2021-07-13 at 3 48 58 PM

Old wizard code is marked as 'deprecated' since it will eventually be removed.

## Stand-alone Prompters

Prompters do not need to be used apart of a wizard. All prompter implementations expose a prompt method (asynchronous) that either returns a wizard control signal (retry, back, exit, etc.) or a type T (the user response). Use the function isValidResponse on the return value to verify the user gave us good data.

## Creating a Wizard (A Very Contrived Example)

All wizard start with a shape (or structure, if you prefer) of the output. Generally speaking this is done by describing an interface (or interfaces). Let's create a wizard that prompts the user for a name, then asks for either a favorite animal or a favorite color based on age:

interface FavoriteRequest {
    name: string
    age: number
    favoriteAnimal?: Animal
    favoriteColor?: {
        isCool: boolean
        color: string
    }
}

interface Animal {
    name: string
    kind: 'Terrestrial' | 'Aquatic' | 'Amphibian'
}

For this example, we will ask for a favorite animal if their age is greater than or equal to 30, otherwise we will ask for a favorite color. To make things interesting, we will also include extra data about the color: whether or not it is 'warm' or 'cool'. The prompt will also let the user enter in any color they want, which means we'll need to ask them whether the custom color is 'warm' or 'cool'. We can describe this form object as follows:

class FavoriteRequestForm extends WizardForm<FavoriteRequest> {
    constructor() {
        super()
        const form = this.body
    
        form.name.bindPrompter(() => createInputBox({ title: 'Enter a name' }))

        // convert string to number (normally we would validate this as well)
        form.age.bindPrompter(state => 
            createInputBox({ title: `Enter ${state.name}'s age` }).transform(resp => Number(resp))
        )

        const animals: Animal[] = [
            { name: 'Tiger', kind: 'Terrestrial' },
            { name: 'Pink Fairy Armadillo', kind: 'Terrestrial' },
            { name: 'Dolphin', kind: 'Aquatic' },
            { name: 'Mantis Shrimp', kind: 'Aquatic' },
            { name: 'Axolotl', kind: 'Amphibian' }
        ]

        // maps our 'Animal' objects into QuickPickItems usable by `createQuickPick`
        const picks = animals.map(animal => ({
            label: animal.name,
            description: animal.kind,
            data: animal,
        }))

        form.favoriteAnimal.bindPrompter(state => 
            createQuickPick(picks, { title: `Select ${state.name}'s favorite animal` }
        ), { showWhen: state => state.age !== undefined && state.age >= 30 })

        function isCool(color: string): boolean {
            return color === 'Blue' || color === 'Green' || color === 'Purple'
        }

        // converts string to 'Title Case'
        // 'a StRinG' -> 'A String'
        function normalize(resp: string): string {
            return resp.split(' ')
                       .map(word => `${word[0].toUpperCase()}${word.slice(1).toLowerCase()}`)
                       .join(' ')
        }

        const colors = ['Red', 'Blue', 'Green', 'Yellow', 'Orange', 'Purple']

        // we don't need to assign the `data` field since the `label` is the data
        const colorItems = colors.map(color => ({ label: color }))

        // gives some preset colors, and allows the user to enter their own color
        form.favoriteColor.color.bindPrompter(state => createLabelQuickPick(colorItems, { 
            title: `Select ${state.name}'s favorite color`,
            // enables 'custom' input, displaying the given label and applying a transformation function
            customInputSettings: {
                label: 'Custom color',
                transform: resp => normalize(resp)
            }
        }), { showWhen: state => state.age !== undefined && state.age < 30 })

        // if the user entered a custom color, we need to ask whether it's a 'warm' or 'cool' color
        form.favoriteColor.isCool.bindPrompter(state => createQuickPick(
            [{ label: 'Warm', data: false }, { label: 'Cool', data: true }], 
            { title: `Is ${state.favoriteColor.color!} a warm or cool color?`}
        ), {
            // only prompt if the user did _not_ select a color from our preset colors
            showWhen: state => colors.indexOf(state.favoriteColor.color!) === -1,
            // sets the field using `isCool` is we did not prompt at all
            setDefault: state => isCool(state.favoriteColor.color!),
            // this prevents the property from being written to if the 'Animal' path was taken
            requireParent: true 
        })
    }
}

Keep in mind that in practice one would normally extract out the prompter creation into their own functions for reusability/testability. This example does not do that. Creating the wizard and running it is then simple:

const wizard = new Wizard(new FavoriteRequestForm())
// type will automatically be inferred as 'FavoriteRequest'
const response = await wizard.run() 

See the previous PR for help in regards to testing WizardForms.
License

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.