App Runner: wizard implementations #1916
There's some extensions to the wizard framework here, mainly just adding a way to have an 'exit' prompt and functionality to estimate the # of steps that would be needed if the user goes down a certain path.

The majority of this PR is just creating a variety of different prompts and then routing them to the correct location. Due to limited time some things aren't quite how I would like them. For example, many of the very simple prompts are not directly tested (like enter a port, name, etc.).