App Runner: activate #1924
Register the commands + telemetry.
I've added a few small fixes and added QuickPick items for when API calls error out. I'll be adding a generic solution for this later, but for now this will work.

Also I noticed that App Runner can get in a really bizarre state where it will list services but not give any operations for them. I confirmed that this is happening with the console as well. So I've added a check for that in the service nodes. I likely never noticed this before with the batch polling stuff since bad data was discarded implicitly as there would be no nodes listening for it.