Merge feature/apprunner #1922
* Apprunner: add client #1911
Adds the client and bumps SDK version.
* App Runner: wizard implementations #1916 
There's some extensions to the wizard framework here, mainly just adding a way to have an 'exit' prompt and functionality to estimate the # of steps that would be needed if the user goes down a certain path.
The majority of this PR is just creating a variety of different prompts and then routing them to the correct location. Due to limited time some things aren't quite how I would like them. For example, many of the very simple prompts are not directly tested (like enter a port, name, etc.).
* App Runner: add nodes to explorer tree #1915 
Adds nodes for App Runner.
The CloudWatch log group node was slightly restructured to have a base class so any resource node could potentially have log nodes. This should be reworked to not use inheritance but for this simple use-case it easiest to do it like that.
* App Runner: activate #1924 
Register the commands + telemetry.
I've added a few small fixes and added QuickPick items for when API calls error out. I'll be adding a generic solution for this later, but for now this will work.
Also I noticed that App Runner can get in a really bizarre state where it will list services but not give any operations for them. I confirmed that this is happening with the console as well. So I've added a check for that in the service nodes. I likely never noticed this before with the batch polling stuff since bad data was discarded implicitly as there would be no nodes listening for it.