SAM/CFN JSON schema: write downloaded schema from memory #1936
## Problem
Whenever the schemas need to be downloaded (i.e. first-time load or an update) they are currently being written to disk with a writestream. Normally not a big deal, however, the YAML extension may try to load the partial schema, requiring a reload to fix things.

## Solution
Write to disk after retrieving the full contents in memory (only for schemas, write streams are normally a good thing)