credentials: let SDK parse credentials if not using 'source_profile' #1953
Problem
-------
Migrating to SDK v3 introduced a regression where credentials were not updated
as the credentials file changed. This was because the new 'patching' logic
assumed the profile data within the provider was accurate (it was not).

We should look towards some simple integration tests for credentials. This
would also allow 'safer' updates to the system in the future.

Solution
--------
Let the SDK parse the file when `source_profile` is not being used, exerting
the original behavior.

Note that the original behavior is likely an 'accident' and highly unintuitive.
The provider class contains information about the shared-ini files that is no
longer in-sync with what the SDK will read. But since it does have the desired
effect I suppose it falls into 'it's not a bug, it's a feature'