SAM/CFN JSON schema: use YAML ext API, rework schema mapping
Problem: The current implementation for YAML schema mapping and activation has a number of drawbacks. Mapping of file to schema is stored in global extension configuration, which is messy and provides no way to clean orphaned file entries. Race condition bugs exist causing mappings to not be stored consistently. Schema versions are also not properly cached, leading to unnecessary redownloads on each activation.
Solution: This change introduces a schema service responsible for managing schema mappings during the lifecycle of the extension, and directly interfaces with the YAML extension to remove the need to write mappings to the extension configuration. The service processes schema mapping updates in batch and handles the cases of schemas or the YAML extension not being available and enforces consistent write behavior. Schemas now properly cache their version on download.