build: move to Vue3 #1995
Moves to Vue3, and refactors the SAM invoke webview as a Vue single-file component (SFC)
This embeds the template + front-end logic + style in a single file. This isn't a requirement, however, it does make sense considering these 3 things are very closely related. The benefit of using the SFC is syntax hightlighting + auto-complete + editor type-safety.