build: bump TypeScript to 4.4.2 #2010
Bump TS to 4.4.2

Errors in catch blocks are now typed as unknown by default (this is the 'strict' setting since it's the safest)
Also we have quite a few catch blocks that just catch errors, log them (with no other context...) and throw them again. This basically makes the error messages useless in production since the stack traces will contain mangled function names and all refer to the bundled script. Re-throwing also means that something else may log the error, so we'll get duplicate messages too.

Best practice: don't catch/throw an error unless you intend to do something with it other than logging. Checking for error codes to react to within the local context then re-throwing is OK, but writing the error to the logs with no context is noise.