credentials: SSO reconnects if valid
Problem: SSO profiles do not reconnect automatically on start up, even
if the token is still valid.
Solution: If the most recent credentials were a SSO profile, check if
the token is still valid. Continue login if it is, but do not attempt to
refresh if the token is no longer valid.