credentials: change watcher type, watch config
Problem: Previous watcher was fs.watchFile. This method polls for
changes which uses the CPU, even when no changes are made. Also, the
config file was not being watched, which can contain full credentials.
Solution: Switch to fs.watch. No polling, listens for a change event.
This also can watch a directory and the event will pass the filename.