MDE create/connect from URI
Adds MDE create and connect feature from URI
Currently creation and connection is just on a single path /mde. Since this is supposed to be for CAWS it'll most likely get moved to the CAWS module and split up. Do we need to support reconnecting to an MDE from the browser? CAWS has APIs for managing MDEs so I'm guessing it's pretty likely.

How to use:
Paste this into your browser to create an MDE: vscode://amazonwebservices.aws-toolkit-vscode/mde
Make sure to select the correct target window
To connect to an already existing MDE:
vscode://amazonwebservices.aws-toolkit-vscode/mde?id=[MDE_ID_HERE]

A progress notification is shown while waiting for an MDE to start as well. The notification should probably be parameterized since the MDE is being started with the context of connecting to it. We should not have a notification when a user starts an MDE from the explorer.