remote invoke: file select from workspace (#2009)
* remote invoke: file select from workspace

Problem: 'Inovke on AWS' form for remote invoke has the option to select
a file to use as the payload. This is opens a file selection window. In
VSC, this should be your local workspace, but in Cloud9, it should be
from the IDE's workspace.
Solution:  Switch from the HTML file input to use the VSC API to show
the open dialog.

* style: improve spacing

* add try/catch for readFile