feat: dynamic resource support in AWS Explorer (#2129)
Problem
-------
Many resource types are currently unsupported in the AWS Explorer. This forces users to use the console or other tools to discover or check statuses for these resources.

Solution
--------
This change introduces dynamic support for discovering and interacting with resources in the explorer. CloudControl APIs are used to generically interact with supported resource types.