build(deps): unwrap UMD imports for JSON language service #2141
The vscode-json-languageservice module includes ESM and UMD emits, but sets the UMD output as the main one. Our webpack config needs extra configuration to ignore UMD distributions to avoid issues like this, but for now we can use the compatibility loader to unwrap the UMD import as we're already using it for the ASL server.