test(UI): add quick pick tester #2168
Adds a testing construct for easy testing of quick picks. Tests are constructed as a series of 'actions' that are executed sequentially. Supports both 'partial' assertions and exact matches. Also adds custom error messages for quicker debugging when tests break.

Some of the logic is not being used currently but that's just because we haven't written many picker tests. There's also partial support for testing multi-pick pickers but of course those aren't even implemented yet. Later on the test code for QuickPickPrompter will use this tester instead, at least for some of the logic.

Added documentation