cloudControl: special-case S3::Bucket to regionalize bucket resolution
Problem: S3::Bucket's resource handler LIST operation returns buckets from all regions. This is inconsistent with all other resource types which return regionalized results and results in an account wide listing of buckets being presented to the user. Buckets out of the current node's region are not accessible since the subsequent GET operation is region scoped.
Fix: This change special cases S3::Bucket to use the regionalized listBuckets function on the toolkit's S3Client to scope the buckets that are resolved to the current node's region.