ECS credentials: handle no attached role
Problem: Previously the Toolkit checked for the availability of ECS container credentials purely on the presence of ECS environment variables. This check does not guarantee that the container has credentials and subsequently, if no role was attached to the ECS container, an exception would be thrown when trying to use the provider.
Solution: If ECS environment variables are present, attempt to retrieve credentials during the provider availability check to validate that a role is attached and credentials are available before making this provider available to the user.