Logs: remove/change some noisy logging code #2274
Fetchers log the error as an error and then return undefined. If something can be handled by just returning undefined then it definitely wasn't an error we really cared about. Also, we should refrain from just dumping the whole stack trace every time an error occurs. It better be an error that truly was unhandled to dump the stack. Especially since the names are mangled in the bundled extension, so it's pretty much useless in production too.