experiments: add UI for configuring experiments
Problem
Experiments can currently only be enabled via the extension configuration. This hurts discoverability and makes it harder for users to configure experiments.

Solution
This change adds a command which exposes a multi select quick pick for enabling experiments. This is exposed on a context menu on the top level AWS Explorer as well as in the command palette.