util: introduce AsyncCollection (#2402)
To make it easier to handle pagination, I've added a new utility interface called AsyncCollection. This basically allows you to treat async generators/iterables as if they were a static collection. This is great since we can manipulate them in a declarative way; no more while loops to iterate over the elements and check for done. It also helps to abstract away async generators, which most callers don't really care about. You can still iterate over the collection as if it were an AsyncIterable, but again the common-case is we just want all the results. This is provided by a promise method which outputs all results in an array.

Benefits:
* Much better ergonomics and type safety for mapping paginated API calls
* Flexible. The collection does not hide the fact that things come in pages, so it can be used in ChildNodeLoader
* Able to map generators themselves, which is important when generators have special meaning behind their 'return' value.
* Collections can be passed around and recognized as AsyncIterable, hiding implementation details